update: dp2px return float

+++ androidktx/src/main/java/com/dengzii/ktx/android/AndroidExt.kt
@@ -6,8 +6,8 @@ fun Float.px2dp(): Float {
     return (this / getDeviceDensity())
 }
 
-fun Float.dp2px(): Int {
-    return (0.5f + this * getDeviceDensity()).toInt()
+fun Float.dp2px(): Float {
+    return (0.5f + this * getDeviceDensity())
 }
 
 private fun getDeviceDensity(): Float {
